商品服务-API-新增商品-商品新增业务流程分析

+++ gulimall-product/src/main/java/com/cr/gulimall/product/vo/Attr.java
@@ -13,7 +13,7 @@ import lombok.Data;
 @Data
 public class Attr {
 
-    private int attrId;
+    private Long attrId;
 
     private String attrName;
